feat: migrate /messages endpoints to MessageController
