+++ src/main/java/br/com/conectabyte/profissu/services/ConversationService.java
@@ -4,13 +4,10 @@ import java.util.List;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.stereotype.Service;
 
 import br.com.conectabyte.profissu.dtos.request.ConversationRequestDto;
-import br.com.conectabyte.profissu.dtos.request.MessageRequestDto;
 import br.com.conectabyte.profissu.dtos.response.ConversationResponseDto;
-import br.com.conectabyte.profissu.dtos.response.MessageResponseDto;
 import br.com.conectabyte.profissu.entities.Conversation;
 import br.com.conectabyte.profissu.entities.Message;
 import br.com.conectabyte.profissu.entities.RequestedService;
@@ -20,9 +17,7 @@ import br.com.conectabyte.profissu.enums.RequestedServiceStatusEnum;
 import br.com.conectabyte.profissu.exceptions.ResourceNotFoundException;
 import br.com.conectabyte.profissu.exceptions.ValidationException;
 import br.com.conectabyte.profissu.mappers.ConversationMapper;
-import br.com.conectabyte.profissu.mappers.MessageMapper;
 import br.com.conectabyte.profissu.repositories.ConversationRepository;
-import br.com.conectabyte.profissu.repositories.MessageRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 
@@ -30,14 +25,11 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class ConversationService {
   private final ConversationRepository conversationRepository;
-  private final MessageRepository messageRepository;
   private final RequestedServiceService requestedServiceService;
   private final JwtService jwtService;
   private final UserService userService;
-  private final SimpMessagingTemplate simpMessagingTemplate;
 
   private final ConversationMapper conversationMapper = ConversationMapper.INSTANCE;
-  private final MessageMapper messageMapper = MessageMapper.INSTANCE;
 
   public Conversation findById(Long id) {
     final var optionalConversation = conversationRepository.findById(id);
@@ -109,37 +101,6 @@ public class ConversationService {
     conversation.setOfferStatus(offerStatus);
 
     return conversationMapper.conversationToConversationResponseDto(conversationRepository.save(conversation));
-  }
-
-  @Transactional
-  public Page<MessageResponseDto> listMessages(Long id, Pageable pageable) {
-    final var messages = messageRepository.listMessages(id, pageable);
-    return messageMapper.messagePageToMessageResponseDtoPage(messages);
-  }
-
-  @Transactional
-  public MessageResponseDto sendMessage(Long id, MessageRequestDto messageRequestDto) {
-    final var conversation = this.findById(id);
-
-    if (conversation.getOfferStatus() != OfferStatusEnum.PENDING
-        && conversation.getOfferStatus() != OfferStatusEnum.ACCEPTED) {
-      throw new ValidationException("This offer has already been canceled or rejected.");
-    }
-
-    final var userId = this.jwtService.getClaims()
-        .map(claims -> Long.valueOf(claims.get("sub").toString()))
-        .orElseThrow();
-    final var user = userService.findById(userId);
-    final var message = Message.builder()
-        .message(messageRequestDto.message())
-        .conversation(conversation)
-        .user(user)
-        .build();
-    final var messageResponseDto = messageMapper.messageToMessageResponseDto(messageRepository.save(message));
-
-    simpMessagingTemplate.convertAndSend("/topic/conversations/" + id + "/messages", messageResponseDto);
-
-    return messageResponseDto;
   }
 
   private void validate(RequestedService requestedService, User serviceProvider, boolean alreadySubmittedAnOffer) {
