+++ src/test/java/br/com/conectabyte/profissu/controllers/AddressControllerTest.java
@@ -50,8 +50,6 @@ class AddressControllerTest {
   private ObjectMapper objectMapper;
 
   private final AddressMapper addressMapper = AddressMapper.INSTANCE;
-  private final Long userId = 1L;
-  private final Long addressId = 1L;
   private final Address address = AddressUtils.create(UserUtils.create());
   private final AddressRequestDto validRequest = addressMapper.addressToAddressRequestDto(address);
   private final AddressResponseDto responseDto = addressMapper.addressToAddressResponseDto(address);
@@ -62,7 +60,8 @@ class AddressControllerTest {
     when(addressService.register(any(), any())).thenReturn(responseDto);
     when(securityService.isOwner(any())).thenReturn(true);
 
-    mockMvc.perform(post("/addresses?userId={userId}", userId)
+    mockMvc.perform(post("/addresses")
+        .param("userId", "1")
         .contentType(MediaType.APPLICATION_JSON)
         .content(objectMapper.writeValueAsString(validRequest)))
         .andExpect(status().isCreated())
@@ -74,7 +73,8 @@ class AddressControllerTest {
   void shouldReturnBadRequestWhenRequestIsInvalid() throws Exception {
     final var invalidRequest = new AddressRequestDto("", "", "", "", "invalid");
 
-    mockMvc.perform(post("/addresses?userId={userId}", userId)
+    mockMvc.perform(post("/addresses")
+        .param("userId", "1")
         .contentType(MediaType.APPLICATION_JSON)
         .content(objectMapper.writeValueAsString(invalidRequest)))
         .andExpect(status().isBadRequest())
@@ -83,7 +83,7 @@ class AddressControllerTest {
 
   @Test
   void shouldReturnUnauthorizedWhenUserIsNotAuthenticated() throws Exception {
-    mockMvc.perform(post("/addresses/" + userId)
+    mockMvc.perform(post("/addresses/1")
         .contentType(MediaType.APPLICATION_JSON)
         .content(objectMapper.writeValueAsString(validRequest)))
         .andExpect(status().isUnauthorized());
@@ -92,7 +92,8 @@ class AddressControllerTest {
   @Test
   @WithMockUser
   void shouldReturnForbiddenWhenUserHasNoPermission() throws Exception {
-    mockMvc.perform(post("/addresses?userId={userId}", userId)
+    mockMvc.perform(post("/addresses")
+        .param("userId", "1")
         .contentType(MediaType.APPLICATION_JSON)
         .content(objectMapper.writeValueAsString(validRequest)))
         .andExpect(status().isForbidden())
@@ -105,7 +106,7 @@ class AddressControllerTest {
     when(addressService.update(any(), any())).thenReturn(responseDto);
     when(securityAddressService.ownershipCheck(any())).thenReturn(true);
 
-    mockMvc.perform(put("/addresses/" + addressId)
+    mockMvc.perform(put("/addresses/1")
         .contentType(MediaType.APPLICATION_JSON)
         .content(objectMapper.writeValueAsString(validRequest)))
         .andExpect(status().isOk())
@@ -118,7 +119,7 @@ class AddressControllerTest {
     when(addressService.update(any(), any())).thenThrow(new ResourceNotFoundException("Address not found"));
     when(securityAddressService.ownershipCheck(any())).thenReturn(true);
 
-    mockMvc.perform(put("/addresses/" + addressId)
+    mockMvc.perform(put("/addresses/1")
         .contentType(MediaType.APPLICATION_JSON)
         .content(objectMapper.writeValueAsString(validRequest)))
         .andExpect(status().isNotFound())
@@ -128,7 +129,8 @@ class AddressControllerTest {
   @Test
   @WithMockUser
   void shouldReturnBadRequestForMalformedJson() throws Exception {
-    mockMvc.perform(post("/addresses?userId={userId}", userId)
+    mockMvc.perform(post("/addresses")
+        .param("userId", "1")
         .contentType(MediaType.APPLICATION_JSON)
         .content("{invalidJson}"))
         .andExpect(status().isBadRequest())
